pyproject.toml
[project]
name = "api-orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinayak Das",email = "vindas@cisco.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi (==0.104.1)",
    "uvicorn (==0.24.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langchain-openai (>=0.3.25,<0.4.0)",
    "openapi-spec-validator (==0.6.0)",
    "prance (==23.6.21.0)",
    "httpx (==0.25.2)",
    "redis (==5.0.1)",
    "networkx (==3.2.1)",
    "pydantic",
    "pydantic-settings",
    "structlog (==23.2.0)",
    "tenacity (==8.2.3)",
    "python-multipart (==0.0.6)",
    "pyyaml (==6.0.1)",
    "jsonpath-ng (==1.6.0)",
    "python-dotenv (==1.0.0)",
    "py-openapi-schema-to-json-schema (==0.0.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "api-orchestrator"
version = "0.1.0"
description = ""
authors = ["Vinayak Das <vindas@cisco.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
